+++ src/web_ui/src/components/UploadMaterialFlow.jsx
@@ -7,7 +7,7 @@ import FindingCard from "./FindingCard.jsx";
 import MemoryCandidateCard from "./MemoryCandidateCard.jsx";
 import ActionCandidateItem from "./ActionCandidateItem.jsx";
 import { fetchQuestionSession } from "../lib/demoApi.js";
-import { runHermesJob } from "../lib/hermesJobs.js";
+import { runHermesJob } from "../lib/demoJobs.js";
 import useAsyncData from "../lib/useAsyncData.js";
 
 const STEPS = [