+++ src/web_ui/src/views/LearningContentView.jsx
@@ -5,7 +5,7 @@ import SubjectFilter, {
 } from "../components/SubjectFilter.jsx";
 import JobTrigger from "../components/JobTrigger.jsx";
 import { fetchTextbookSummary } from "../lib/demoApi.js";
-import { runHermesJob } from "../lib/hermesJobs.js";
+import { runHermesJob } from "../lib/demoJobs.js";
 import { defaultSubjects, demoTextbookIds } from "../lib/demoData.js";
 import useAsyncData from "../lib/useAsyncData.js";